+++ angular-workspace/projects/ni/nimble-angular/src/test.ts
@@ -18,7 +18,8 @@ declare const require: {
 // First, initialize the Angular testing environment.
 getTestBed().initTestEnvironment(
     BrowserDynamicTestingModule,
-    platformBrowserDynamicTesting(), {
+    platformBrowserDynamicTesting(),
+    {
         teardown: { destroyAfterEach: false }
     }
 );