+++ packages/nimble-components/src/table/index.ts
@@ -1,4 +1,4 @@
-import { observable } from '@microsoft/fast-element';
+import { attr, observable } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import {
     ColumnDef as TanStackColumnDef,
@@ -9,9 +9,10 @@ import {
     getCoreRowModel as tanStackGetCoreRowModel,
     TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
+import { TableValidator } from './models/table-validator';
 import { styles } from './styles';
 import { template } from './template';
-import type { TableRecord } from './types';
+import type { TableRecord, TableRowState, TableValidity } from './types';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -25,8 +26,17 @@ declare global {
 export class Table<
     TData extends TableRecord = TableRecord
 > extends FoundationElement {
+    @attr({ attribute: 'id-field-name' })
+    public idFieldName?: string | null;
+
     @observable
     public data: TData[] = [];
+
+    /**
+     * @internal
+     */
+    @observable
+    public tableData: TableRowState<TData>[] = [];
 
     // TODO: Temporarily expose the columns as a string array. This will ultimately be
     // column definitions provided by slotted elements.
@@ -37,9 +47,14 @@ export class Table<
     @observable
     public columnHeaders: string[] = [];
 
+    public get validity(): TableValidity {
+        return this.tableValidator.getValidity();
+    }
+
     private readonly table: TanStackTable<TData>;
     private options: TanStackTableOptionsResolved<TData>;
     private readonly tableInitialized: boolean = false;
+    private readonly tableValidator = new TableValidator();
 
     public constructor() {
         super();
@@ -47,6 +62,13 @@ export class Table<
             data: [],
             onStateChange: (_: TanStackUpdater<TanStackTableState>) => {},
             getCoreRowModel: tanStackGetCoreRowModel(),
+            getRowId: record => {
+                if (this.idFieldName) {
+                    return record[this.idFieldName] as string;
+                }
+                // Return a falsey value to use the default ID from TanStack.
+                return '';
+            },
             columns: [],
             state: {},
             renderFallbackValue: null,
@@ -54,6 +76,15 @@ export class Table<
         };
         this.table = tanStackCreateTable(this.options);
         this.tableInitialized = true;
+    }
+
+    public idFieldNameChanged(
+        _prev: string | undefined,
+        _next: string | undefined
+    ): void {
+        // Force TanStack to detect a data update because a row's ID is only
+        // generated when creating a new row model.
+        this.trySetData([...this.data]);
     }
 
     public dataChanged(
@@ -66,8 +97,32 @@ export class Table<
 
         // Ignore any updates that occur prior to the TanStack table being initialized.
         if (this.tableInitialized) {
-            this.updateTableOptions({ data: this.data });
+            this.trySetData(this.data);
         }
+    }
+
+    public checkValidity(): boolean {
+        return this.tableValidator.isValid();
+    }
+
+    private trySetData(newData: TData[]): void {
+        const areIdsValid = this.tableValidator.validateDataIds(
+            newData,
+            this.idFieldName
+        );
+        if (areIdsValid) {
+            this.updateTableOptions({ data: newData });
+        } else {
+            this.updateTableOptions({ data: [] });
+        }
+    }
+
+    private refreshRows(): void {
+        const rows = this.table.getRowModel().rows;
+        this.tableData = rows.map(row => {
+            const rowState: TableRowState<TData> = { data: row.original };
+            return rowState;
+        });
     }
 
     private updateTableOptions(
@@ -75,6 +130,7 @@ export class Table<
     ): void {
         this.options = { ...this.options, ...updatedOptions };
         this.update(this.table.initialState);
+        this.refreshRows();
     }
 
     private readonly update = (state: TanStackTableState): void => {