+++ packages/nimble-components/src/theme-provider/design-token-comments.ts
@@ -29,7 +29,6 @@ export const comments: { readonly [key in TokenName]: string | null } = {
     borderHoverColor: 'Border color when hovering component',
     iconColor: 'Equivalent to the font color for icons',
     modalBackdropColor: 'Color of background overlay behind modal dialog boxes',
-    popupBoxShadowColor: 'Shadow color for menus and dialog boxes',
     popupBorderColor: 'Border color for menus and dialog boxes',
     controlHeight:
         'Standard layout height for all controls. Add "labelHeight" for labels on top.',
@@ -253,5 +252,11 @@ export const comments: { readonly [key in TokenName]: string | null } = {
         'Font line height for the "Error_LightUi" base token',
     errorTextFallbackFontFamily:
         'Fallback font family for the "Error_LightUi" base token',
-    tableRowBorderColor: 'Color for the border of rows in the table'
+    tableRowBorderColor: 'Color for the border of rows in the table',
+    elevation1BoxShadow:
+        'The box shadow for elevation 1. Used for component hover states.',
+    elevation2BoxShadow:
+        'The box shadow for elevation 2. Used for components such as menus, banners, tooltips, error notifications, and scrolling.',
+    elevation3BoxShadow:
+        'The box shadow for elevation 3. Used for components such as dialogs, overlays, and pop-ups.'
 };